Restructured application
- Created /CreateJob/
- Moved files to new folder and renamed MVC to CreateJobModel, CreateJobView, CreateJobController
- Added /User/UserModel.js
- Added AUthEndpoint.txt
- Added UQTheme.zip
