+++ App/styles.css
@@ -27,6 +27,25 @@ h1 {
     font-size: 3.2rem;
 }
 
+/**************
+== Page Grid ==
+***************/
+body {
+    display: grid;
+    grid-template-columns: 50% 50%;
+}
+    .global-header {
+        grid-column: 1 / 3;
+    }
+    .global-navigation {
+        grid-column: 1 / 3;
+    }
+    main {
+        grid-column: 1 / 3;
+    }
+    aside {
+        grid-column: 1 / 3;
+    }
 /**************************
 == Header and Navigation ==
 ***************************/
@@ -34,10 +53,6 @@ h1 {
     background-color: #51247A;
     color: #ccc;
     padding: 1rem;
-}
-
-.gloabl-navigation {
-
 }
     .global-navigation__menu {
         list-style: none;
@@ -79,8 +94,33 @@ main {
         color: #fff;
         background-color: #4bb543;
 
-        font-style: bold;
+        /*font-style: bold;*/
         font-size: 1rem;
         font-weight: 500;
         line-height: 24px;
+    }
+
+
+aside {
+    padding: 1rem;
+    margin: 1rem;
+    background-color: #eeeeee;
+}
+    .code {
+        padding: 1rem;
+        border: 1px dotted lightblue;
+        background-color: white;
+        max-width: 32rem;
+    
+        display: grid;
+        grid-template-columns: auto;
+    }
+    .code__line {
+        margin: 0;
+        
+        display: flex;
+        justify-content: space-between;
+    }
+    .code__line--explain {
+        background-color: yellow;
     }
